+++ Grocery.cpp
@@ -14,7 +14,7 @@
 
 using namespace std;
 
-Grocery::Grocery(string name, float price, unsigned int amount) : Item(name,price,amount) {}
+Grocery::Grocery(string name, double price, unsigned int amount) : Item(name,price,amount) {}
 
 Grocery::~Grocery() {}
